+++ dft2lnt/dft2lnt.h
@@ -17,14 +17,14 @@ class DFT2LNT {
 public:
 	static int validateCommands(const std::vector<DFT::AST::ASTNode*>* ast) {
 		int valid = true;
-		for(int i=0; i<ast->size(); ++i) {
+		for(size_t i=0; i<ast->size(); ++i) {
 //			valid = ast->at(i)->validate() ? valid : false;
 		}
 		return valid;
 	}
 	static DFTree* createDFT(const std::vector<DFT::AST::ASTNode*>* ast) {
 		DFTree* dft = new DFTree();
-		for(int i=0; i<ast->size(); ++i) {
+		for(size_t i=0; i<ast->size(); ++i) {
 //			ast->at(i)->handleCommand(dft);
 		}
 		return dft;
