+++ dft2lnt/dftnodes/Gate.h
@@ -32,7 +32,7 @@ public:
 	virtual ~Gate() {
 	}
 	virtual void addReferencesTo(std::vector<Node*>& nodeList) {
-		for(int i=0; i<children.size(); ++i) {
+		for(size_t i=0; i<children.size(); ++i) {
 			nodeList.push_back(children.at(i));
 		}
 	}
